MainCustomerMenu added the switch for all options. Implemented the
logout option the users Cart gets saved in the DB at logout. Partial
implementation of option "9" in menu: "Naar Kassa / Winkelwagen
wijzigen". Created CartMainMenu class and created the options that
display below the displayed cart. The menu adapts to the number of
subOrders in the cart.

The displayCart method is the same as in the MainCustomerMenu, so I
created an Interface DisplayCart with a default method DisplayCart()